fix: 새로 고침할 경우 북마크 비어지는 버그 수정
